+++ wallet-sidebar.js
@@ -163,6 +163,11 @@ const sidebar = {
               type: "doc",
               label: "Add a network",
               id: "how-to/manage-networks/add-network"
+            },
+            {
+              type: "doc",
+              label: "Interact with multiple networks",
+              id: "how-to/manage-networks/use-multichain"
             }
           ]
         },
@@ -379,6 +384,11 @@ const sidebar = {
         },
         {
           type: "doc",
+          label: "About the Multichain API",
+          id: "concepts/multichain-api"
+        },
+        {
+          type: "doc",
           label: "SDK connections",
           id: "concepts/sdk-connections"
         },
@@ -477,6 +487,11 @@ const sidebar = {
           type: "doc",
           label: "Ethereum provider API",
           id: "reference/provider-api",
+        },
+        {
+          type: "doc",
+          label: "Multichain API",
+          id: "reference/multichain-api",
         },
         {
           type: "category",